+++ src/helpers/index.ts
@@ -33,5 +33,6 @@ export * from './openFileInEditor';
 export * from './parseWinPath';
 export * from './processArticlePlaceholders';
 export * from './processFmPlaceholders';
+export * from './processI18nPlaceholders';
 export * from './processPathPlaceholders';
 export * from './processTimePlaceholders';